requirements.txt
# ===== CORE PYTHON LIBRARIES =====
numpy>=1.23
pandas>=1.5
scipy>=1.10
pyyaml>=6.0

# ===== RL, DEEP LEARNING =====
torch>=2.0
gymnasium==0.29.1       # or gym==0.26 if using older stable-baselines3
stable-baselines3>=2.2.1

# ===== DATA HANDLING/SCALING =====
scikit-learn>=1.2
matplotlib>=3.7
seaborn>=0.12

# ===== OPTIMIZATION / TUNING =====
optuna>=3.5

# ===== LOGGING, UTILS =====
nptyping>=2.5
tqdm>=4.65

# ===== PRODUCTION, SCHEDULING =====
APScheduler>=3.10

# ===== BROKER/EXCHANGE APIs (Uncomment as needed) =====
# smartapi-python>=1.0.2
# angel-one-api>=1.0.0
# kiteconnect>=4.1.0

# ===== DEVELOPMENT/DEBUGGING =====
ipython
jupyter

# ===== CSV/FILE UTILS =====
python-dotenv>=1.0
